+++ software/altimeter/libraries/ArduinoJson/extras/tests/Misc/conflicts.cpp
@@ -1,5 +1,5 @@
 // ArduinoJson - https://arduinojson.org
-// Copyright © 2014-2023, Benoit BLANCHON
+// Copyright © 2014-2024, Benoit BLANCHON
 // MIT License
 
 // Include any header that might use the conflicting macros
@@ -54,6 +54,9 @@
 
 // issue #1905
 #define _current
+
+// issue #1914
+#define V7 7
 
 // catch.hpp mutes several warnings, this file also allows to detect them
 #include "ArduinoJson.h"