+++ software/altimeter/libraries/MCP4725/MCP4725.cpp
@@ -2,10 +2,8 @@
 //    FILE: MCP4725.cpp
 //  AUTHOR: Rob Tillaart
 // PURPOSE: Arduino library for 12 bit I2C DAC - MCP4725
-// VERSION: 0.3.5
+// VERSION: 0.4.0
 //     URL: https://github.com/RobTillaart/MCP4725
-//
-//  HISTORY see changelog.md
 
 
 #include "MCP4725.h"
@@ -27,52 +25,13 @@ MCP4725::MCP4725(const uint8_t deviceAddress, TwoWire *wire)
   _lastValue       = 0;
   _powerDownMode   = 0;
   _lastWriteEEPROM = 0;
+  _maxVoltage      = 5.0;
 }
-
-
-#if defined(ESP8266) || defined(ESP32)
-bool MCP4725::begin(const uint8_t dataPin, const uint8_t clockPin)
-{
-  _wire  = &Wire;
-  if ((dataPin < 255) && (clockPin < 255))
-  {
-    _wire->begin(dataPin, clockPin);
-  } else {
-    _wire->begin();
-  }
-  if (isConnected())
-  {
-    _lastValue = readDAC();
-    _powerDownMode = readPowerDownModeDAC();
-    return true;
-  }
-  return false;
-}
-#endif
-
-#if defined (ARDUINO_ARCH_RP2040)
-
-bool MCP4725::begin(int sda, int scl)
-{
-  _wire->setSDA(sda);
-  _wire->setSCL(scl);
-  _wire->begin();
-
-  if (isConnected())
-  {
-    _lastValue = readDAC();
-    _powerDownMode = readPowerDownModeDAC();
-    return true;
-  }
-  return false;
-}
-
-#endif
 
 
 bool MCP4725::begin()
 {
-  _wire->begin();
+  if ((_deviceAddress < 0x60) || (_deviceAddress > 0x67)) return false;
   if (! isConnected()) return false;
 
   _lastValue = readDAC();
@@ -88,9 +47,15 @@ bool MCP4725::isConnected()
 }
 
 
+uint8_t MCP4725::getAddress()
+{
+  return _deviceAddress;
+}
+
+
 int MCP4725::setValue(const uint16_t value)
 {
-  if (value == _lastValue) return 0;
+  if (value == _lastValue) return MCP4725_OK;
   if (value > MCP4725_MAXVALUE) return MCP4725_VALUE_ERROR;
   int rv = _writeFastMode(value);
   if (rv == 0) _lastValue = value;
@@ -111,6 +76,36 @@ int MCP4725::setPercentage(float percentage)
 }
 
 
+float MCP4725::getPercentage()
+{
+  return getValue() * (100.0 / MCP4725_MAXVALUE);
+}
+
+
+void MCP4725::setMaxVoltage(float v)
+{
+  _maxVoltage = v;
+}
+
+
+float MCP4725::getMaxVoltage()
+{
+  return _maxVoltage;
+}
+
+
+int MCP4725::setVoltage(float v)
+{
+  return setValue(round((v * MCP4725_MAXVALUE) / _maxVoltage));
+}
+
+
+float MCP4725::getVoltage()
+{
+  return getValue() * (_maxVoltage / MCP4725_MAXVALUE);
+}
+
+
 //  unfortunately it is not possible to write a different value
 //  to the DAC and EEPROM simultaneously or write EEPROM only.
 int MCP4725::writeDAC(const uint16_t value, const bool EEPROM)
@@ -120,6 +115,17 @@ int MCP4725::writeDAC(const uint16_t value, const bool EEPROM)
   int rv = _writeRegisterMode(value, EEPROM ? MCP4725_DACEEPROM : MCP4725_DAC);
   if (rv == 0) _lastValue = value;
   return rv;
+}
+
+
+//  ready checks if the last write to EEPROM has been written.
+//  until ready all writes to the MCP4725 are ignored!
+bool MCP4725::ready()
+{
+  yield();
+  uint8_t buffer[1];
+  _readRegister(buffer, 1);
+  return ((buffer[0] & 0x80) > 0);
 }
 
 
@@ -147,6 +153,12 @@ uint16_t MCP4725::readEEPROM()
 }
 
 
+uint32_t MCP4725::getLastWriteEEPROM()
+{
+  return _lastWriteEEPROM;
+};
+
+
 //  depending on bool EEPROM the value of PDM is written to
 //  (false) DAC or
 //  (true) DAC & EEPROM,
@@ -169,7 +181,7 @@ uint8_t MCP4725::readPowerDownModeEEPROM()
 
 uint8_t MCP4725::readPowerDownModeDAC()
 {
-  while(!ready());  // TODO needed?
+  while(!ready());  //  TODO needed?
   uint8_t buffer[1];
   _readRegister(buffer, 1);
   uint8_t value = (buffer[0] >> 1) & 0x03;
@@ -183,7 +195,7 @@ uint8_t MCP4725::readPowerDownModeDAC()
 int MCP4725::powerOnReset()
 {
   int rv = _generalCall(MCP4725_GC_RESET);
-  _lastValue = readDAC(); // update cache to actual value;
+  _lastValue = readDAC(); //  update cache to actual value;
   return rv;
 }
 
@@ -194,32 +206,22 @@ int MCP4725::powerOnReset()
 int MCP4725::powerOnWakeUp()
 {
   int rv = _generalCall(MCP4725_GC_WAKEUP);
-  _powerDownMode = readPowerDownModeDAC();  // update to actual value;
+  _powerDownMode = readPowerDownModeDAC();  //  update to actual value;
   return rv;
 }
+
 
 //  PAGE 18 DATASHEET
 int MCP4725::_writeFastMode(const uint16_t value)
 {
   uint8_t l = value & 0xFF;
-  uint8_t h = ((value / 256) & 0x0F);  // set C0 = C1 = 0, no PDmode
+  uint8_t h = ((value / 256) & 0x0F);  //  set C0 = C1 = 0, no PDmode
   h = h | (_powerDownMode << 4);
 
   _wire->beginTransmission(_deviceAddress);
   _wire->write(h);
   _wire->write(l);
   return _wire->endTransmission();
-}
-
-
-//  ready checks if the last write to EEPROM has been written.
-//  until ready all writes to the MCP4725 are ignored!
-bool MCP4725::ready()
-{
-  yield();
-  uint8_t buffer[1];
-  _readRegister(buffer, 1);
-  return ((buffer[0] & 0x80) > 0);
 }
 
 
@@ -248,9 +250,9 @@ uint8_t MCP4725::_readRegister(uint8_t* buffer, const uint8_t length)
 {
   _wire->beginTransmission(_deviceAddress);
   int rv = _wire->endTransmission();
-  if (rv != 0) return 0;  // error
+  if (rv != 0) return 0;  //  error
 
-  //  readbytes will always be equal or smaller to length
+  //  readBytes will always be equal or smaller to length
   uint8_t readBytes = _wire->requestFrom(_deviceAddress, length);
   uint8_t cnt = 0;
   while (cnt < readBytes)
@@ -264,7 +266,7 @@ uint8_t MCP4725::_readRegister(uint8_t* buffer, const uint8_t length)
 //  name comes from datasheet
 int MCP4725::_generalCall(const uint8_t gc)
 {
-  _wire->beginTransmission(0);  // _deviceAddress
+  _wire->beginTransmission(0);  //  _deviceAddress
   _wire->write(gc);
   return _wire->endTransmission();
 }